requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
sqlalchemy==2.0.23
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask==2.2.5
